fix(zigshop): cast long->int pour ItemStack (retrait banque)
ItemStack n'a pas de constructeur (Item, long) ; le net du retrait
(WithdrawResult.net()) est long en interne mais toujours <= amount
(deja un int reseau) donc le cast est sans risque.

+++ zig-shop-mod/src/main/java/com/rawstudio/zigshop/BankServerHandler.java
@@ -164,7 +164,8 @@ public final class BankServerHandler {
             return;
         }
         if (result.net() > 0) {
-            ItemStack out = new ItemStack(currency, result.net());
+            // (int) sans risque : net <= amount, et amount est deja un int (parametre reseau).
+            ItemStack out = new ItemStack(currency, (int) result.net());
             if (!player.getInventory().add(out)) {
                 player.drop(out, false);
             }
